refactor: move issue output contracts to output crate
