+++ crates/cli/src/schema.rs
@@ -2,8 +2,9 @@ use std::process::ExitCode;
 
 use clap::CommandFactory;
 #[cfg(test)]
-use fallow_types::issue_meta::result_issue_metas;
-use fallow_types::issue_meta::{TsAliasMeta, issue_meta_by_code, issue_result_meta_by_code};
+use fallow_output::issue_output_contracts;
+use fallow_output::{TsAliasMeta, issue_output_contract_by_code};
+use fallow_types::issue_meta::issue_meta_by_code;
 use fallow_types::mcp_manifest::{MCP_TOOLS, RUNTIME_COVERAGE_LICENSE_NOTE};
 
 use crate::Cli;
@@ -137,120 +138,19 @@ impl IssueTypeMeta {
                 meta.suppress = Some((token, shared.suppress_file_level));
             }
         }
-        if let Some(result) = issue_result_meta_by_code(bare_id) {
-            meta.result_key = Some(result.result_key);
-            meta.summary_label = issue_summary_label(bare_id);
-            meta.summary_docs_anchor = issue_summary_docs_anchor(bare_id);
-            meta.sarif_rule_ids = Some(result.sarif_rule_ids());
-            let codeclimate_check_names = result.codeclimate_check_names();
-            if !codeclimate_check_names.is_empty() {
-                meta.codeclimate_check_names = Some(codeclimate_check_names);
+        if let Some(contract) = issue_output_contract_by_code(bare_id) {
+            meta.result_key = Some(contract.result_key);
+            meta.summary_label = Some(contract.summary_label);
+            meta.summary_docs_anchor = Some(contract.summary_docs_anchor);
+            meta.sarif_rule_ids = Some(contract.sarif_rule_ids);
+            if !contract.codeclimate_check_names.is_empty() {
+                meta.codeclimate_check_names = Some(contract.codeclimate_check_names);
             }
-            meta.ts_alias = result.ts_alias();
-            meta.counts_in_total = result.counts_in_total;
+            meta.ts_alias = contract.ts_alias;
+            meta.counts_in_total = contract.counts_in_total;
         }
         meta
     }
-}
-
-fn issue_summary_label(code: &str) -> Option<&'static str> {
-    Some(match code {
-        "unused-file" => "Unused files",
-        "unused-export" => "Unused exports",
-        "unused-type" => "Unused types",
-        "private-type-leak" => "Private type leaks",
-        "unused-dependency" => "Unused dependencies",
-        "unused-dev-dependency" => "Unused devDependencies",
-        "unused-optional-dependency" => "Unused optionalDependencies",
-        "unused-enum-member" => "Unused enum members",
-        "unused-class-member" => "Unused class members",
-        "unused-store-member" => "Unused store members",
-        "unresolved-import" => "Unresolved imports",
-        "unlisted-dependency" => "Unlisted dependencies",
-        "duplicate-export" => "Duplicate exports",
-        "type-only-dependency" => "Type-only dependencies",
-        "test-only-dependency" => "Test-only dependencies",
-        "circular-dependency" => "Circular dependencies",
-        "re-export-cycle" => "Re-export cycles",
-        "boundary-violation" => "Boundary violations",
-        "boundary-coverage" => "Boundary coverage",
-        "boundary-call-violation" => "Boundary calls",
-        "policy-violation" => "Policy violations",
-        "invalid-client-export" => "Invalid client exports",
-        "mixed-client-server-barrel" => "Mixed client/server barrels",
-        "misplaced-directive" => "Misplaced directives",
-        "unprovided-inject" => "Unprovided injects",
-        "unrendered-component" => "Unrendered components",
-        "unused-component-prop" => "Unused component props",
-        "unused-component-emit" => "Unused component emits",
-        "unused-component-input" => "Unused component inputs",
-        "unused-component-output" => "Unused component outputs",
-        "unused-svelte-event" => "Unused Svelte events",
-        "unused-server-action" => "Unused server actions",
-        "unused-load-data-key" => "Unused load data keys",
-        "route-collision" => "Route collisions",
-        "dynamic-segment-name-conflict" => "Dynamic segment conflicts",
-        "stale-suppression" => "Stale suppressions",
-        "unused-catalog-entry" => "Unused catalog entries",
-        "empty-catalog-group" => "Empty catalog groups",
-        "unresolved-catalog-reference" => "Unresolved catalog references",
-        "unused-dependency-override" => "Unused dependency overrides",
-        "misconfigured-dependency-override" => "Misconfigured dependency overrides",
-        "prop-drilling" => "Prop drilling",
-        "thin-wrapper" => "Thin wrappers",
-        "duplicate-prop-shape" => "Duplicate prop shapes",
-        _ => return None,
-    })
-}
-
-fn issue_summary_docs_anchor(code: &str) -> Option<&'static str> {
-    Some(match code {
-        "unused-file" => "unused-files",
-        "unused-export" => "unused-exports",
-        "unused-type" => "unused-types",
-        "private-type-leak" => "private-type-leaks",
-        "unused-dependency" | "unused-dev-dependency" | "unused-optional-dependency" => {
-            "unused-dependencies"
-        }
-        "unused-enum-member" => "unused-enum-members",
-        "unused-class-member" => "unused-class-members",
-        "unused-store-member" => "unused-store-members",
-        "unresolved-import" => "unresolved-imports",
-        "unlisted-dependency" => "unlisted-dependencies",
-        "duplicate-export" => "duplicate-exports",
-        "type-only-dependency" => "type-only-dependencies",
-        "test-only-dependency" => "test-only-dependencies",
-        "circular-dependency" => "circular-dependencies",
-        "re-export-cycle" => "re-export-cycles",
-        "boundary-violation" | "boundary-coverage" | "boundary-call-violation" => {
-            "boundary-violations"
-        }
-        "policy-violation" => "policy-violations",
-        "invalid-client-export" => "invalid-client-exports",
-        "mixed-client-server-barrel" => "mixed-client-server-barrels",
-        "misplaced-directive" => "misplaced-directives",
-        "unprovided-inject" => "unprovided-inject",
-        "unrendered-component" => "unrendered-component",
-        "unused-component-prop" => "unused-component-prop",
-        "unused-component-emit" => "unused-component-emit",
-        "unused-component-input" => "unused-component-input",
-        "unused-component-output" => "unused-component-output",
-        "unused-svelte-event" => "unused-svelte-event",
-        "unused-server-action" => "unused-server-action",
-        "unused-load-data-key" => "unused-load-data-key",
-        "dynamic-segment-name-conflict" => "dynamic-segment-name-conflicts",
-        "route-collision" => "route-collisions",
-        "stale-suppression" => "stale-suppressions",
-        "unused-catalog-entry" => "unused-catalog-entries",
-        "empty-catalog-group" => "empty-catalog-groups",
-        "unresolved-catalog-reference" => "unresolved-catalog-references",
-        "unused-dependency-override" => "unused-dependency-overrides",
-        "misconfigured-dependency-override" => "misconfigured-dependency-overrides",
-        "prop-drilling" => "prop-drilling",
-        "thin-wrapper" => "thin-wrapper",
-        "duplicate-prop-shape" => "duplicate-prop-shape",
-        _ => return None,
-    })
 }
 
 fn issue_types_schema() -> serde_json::Value {
@@ -1328,11 +1228,11 @@ mod tests {
                 ))
             })
             .collect();
-        let expected: FxHashSet<(String, String, String)> = result_issue_metas()
-            .filter_map(|meta| {
-                let alias = meta.ts_alias()?;
+        let expected: FxHashSet<(String, String, String)> = issue_output_contracts()
+            .filter_map(|contract| {
+                let alias = contract.ts_alias?;
                 Some((
-                    meta.code.to_string(),
+                    contract.code.to_string(),
                     alias.name.to_string(),
                     alias.parent.to_string(),
                 ))
@@ -1371,7 +1271,7 @@ mod tests {
         )));
         assert_eq!(
             expected, aliases,
-            "schema ts_alias rows must exactly mirror IssueResultMeta::ts_alias()"
+            "schema ts_alias rows must exactly mirror fallow-output contracts"
         );
 
         for (_, name, parent) in aliases {
