+++ crates/types/src/issue_meta.rs
@@ -711,139 +711,6 @@ pub struct IssueResultMeta {
     pub counts_in_total: bool,
 }
 
-/// TypeScript backwards-compat alias emitted for this result row.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct TsAliasMeta {
-    /// Bare alias name kept available from the published `fallow/types` subpath.
-    pub name: &'static str,
-    /// Generated `*Finding` wrapper type the alias resolves to.
-    pub parent: &'static str,
-}
-
-impl IssueResultMeta {
-    /// SARIF rule ids used by the CLI SARIF formatter for this result row.
-    #[must_use]
-    pub fn sarif_rule_ids(self) -> Vec<String> {
-        let mut ids = vec![format!("fallow/{}", self.code)];
-        if self.code == "stale-suppression" {
-            ids.push("fallow/missing-suppression-reason".to_string());
-        }
-        ids
-    }
-
-    /// Whether the CodeClimate formatter emits this result row.
-    #[must_use]
-    pub fn emits_codeclimate(self) -> bool {
-        CODECLIMATE_RESULT_CODES.contains(&self.code)
-    }
-
-    /// CodeClimate check names used by the CLI CodeClimate formatter.
-    #[must_use]
-    pub fn codeclimate_check_names(self) -> Vec<String> {
-        if !self.emits_codeclimate() {
-            return Vec::new();
-        }
-        self.sarif_rule_ids()
-    }
-
-    /// Published TypeScript alias policy for backwards-compatible bare names.
-    #[must_use]
-    pub fn ts_alias(self) -> Option<TsAliasMeta> {
-        let alias = match self.code {
-            "unused-file" => TsAliasMeta {
-                name: "UnusedFile",
-                parent: "UnusedFileFinding",
-            },
-            "unused-export" => TsAliasMeta {
-                name: "UnusedExport",
-                parent: "UnusedExportFinding",
-            },
-            "private-type-leak" => TsAliasMeta {
-                name: "PrivateTypeLeak",
-                parent: "PrivateTypeLeakFinding",
-            },
-            "unused-dependency" => TsAliasMeta {
-                name: "UnusedDependency",
-                parent: "UnusedDependencyFinding",
-            },
-            "unused-dev-dependency" => TsAliasMeta {
-                name: "UnusedDependency",
-                parent: "UnusedDevDependencyFinding",
-            },
-            "unused-optional-dependency" => TsAliasMeta {
-                name: "UnusedDependency",
-                parent: "UnusedOptionalDependencyFinding",
-            },
-            "unused-enum-member" => TsAliasMeta {
-                name: "UnusedMember",
-                parent: "UnusedEnumMemberFinding",
-            },
-            "unused-class-member" => TsAliasMeta {
-                name: "UnusedMember",
-                parent: "UnusedClassMemberFinding",
-            },
-            "unused-store-member" => TsAliasMeta {
-                name: "UnusedMember",
-                parent: "UnusedStoreMemberFinding",
-            },
-            "unresolved-import" => TsAliasMeta {
-                name: "UnresolvedImport",
-                parent: "UnresolvedImportFinding",
-            },
-            "unlisted-dependency" => TsAliasMeta {
-                name: "UnlistedDependency",
-                parent: "UnlistedDependencyFinding",
-            },
-            "duplicate-export" => TsAliasMeta {
-                name: "DuplicateExport",
-                parent: "DuplicateExportFinding",
-            },
-            "type-only-dependency" => TsAliasMeta {
-                name: "TypeOnlyDependency",
-                parent: "TypeOnlyDependencyFinding",
-            },
-            "test-only-dependency" => TsAliasMeta {
-                name: "TestOnlyDependency",
-                parent: "TestOnlyDependencyFinding",
-            },
-            "circular-dependency" => TsAliasMeta {
-                name: "CircularDependency",
-                parent: "CircularDependencyFinding",
-            },
-            "re-export-cycle" => TsAliasMeta {
-                name: "ReExportCycle",
-                parent: "ReExportCycleFinding",
-            },
-            "boundary-violation" => TsAliasMeta {
-                name: "BoundaryViolation",
-                parent: "BoundaryViolationFinding",
-            },
-            "unused-catalog-entry" => TsAliasMeta {
-                name: "UnusedCatalogEntry",
-                parent: "UnusedCatalogEntryFinding",
-            },
-            "empty-catalog-group" => TsAliasMeta {
-                name: "EmptyCatalogGroup",
-                parent: "EmptyCatalogGroupFinding",
-            },
-            "unresolved-catalog-reference" => TsAliasMeta {
-                name: "UnresolvedCatalogReference",
-                parent: "UnresolvedCatalogReferenceFinding",
-            },
-            "unused-dependency-override" => TsAliasMeta {
-                name: "UnusedDependencyOverride",
-                parent: "UnusedDependencyOverrideFinding",
-            },
-            "misconfigured-dependency-override" => TsAliasMeta {
-                name: "MisconfiguredDependencyOverride",
-                parent: "MisconfiguredDependencyOverrideFinding",
-            },
-            _ => return None,
-        };
-        Some(alias)
-    }
-}
-
 /// All shared issue-to-result metadata rows.
 pub const ISSUE_RESULT_META: &[IssueResultMeta] = &[
     IssueResultMeta {
@@ -1066,51 +933,6 @@ pub const ISSUE_RESULT_META: &[IssueResultMeta] = &[
         result_key: "duplicate_prop_shapes",
         counts_in_total: false,
     },
-];
-
-/// Result issue codes emitted by the dead-code CodeClimate formatter.
-pub const CODECLIMATE_RESULT_CODES: &[&str] = &[
-    "unused-file",
-    "unused-export",
-    "unused-type",
-    "private-type-leak",
-    "unused-dependency",
-    "unused-dev-dependency",
-    "unused-optional-dependency",
-    "unused-enum-member",
-    "unused-class-member",
-    "unused-store-member",
-    "unresolved-import",
-    "unlisted-dependency",
-    "duplicate-export",
-    "type-only-dependency",
-    "test-only-dependency",
-    "circular-dependency",
-    "re-export-cycle",
-    "boundary-violation",
-    "boundary-coverage",
-    "boundary-call-violation",
-    "policy-violation",
-    "invalid-client-export",
-    "mixed-client-server-barrel",
-    "misplaced-directive",
-    "unprovided-inject",
-    "unrendered-component",
-    "unused-component-prop",
-    "unused-component-emit",
-    "unused-component-input",
-    "unused-component-output",
-    "unused-svelte-event",
-    "unused-server-action",
-    "unused-load-data-key",
-    "route-collision",
-    "dynamic-segment-name-conflict",
-    "stale-suppression",
-    "unused-catalog-entry",
-    "empty-catalog-group",
-    "unresolved-catalog-reference",
-    "unused-dependency-override",
-    "misconfigured-dependency-override",
 ];
 
 /// Canonical names and aliases accepted by `IssueKind::parse`.
@@ -1399,93 +1221,6 @@ mod tests {
                 meta.result_key
             );
         }
-    }
-
-    #[test]
-    fn result_ci_format_contracts_are_present() {
-        for meta in ISSUE_RESULT_META {
-            let sarif_rule_ids = meta.sarif_rule_ids();
-            assert!(
-                sarif_rule_ids.contains(&format!("fallow/{}", meta.code)),
-                "result metadata code {} has wrong SARIF rule id",
-                meta.code
-            );
-            for rule_id in sarif_rule_ids {
-                assert!(
-                    rule_id.starts_with("fallow/"),
-                    "result metadata code {} has unprefixed SARIF rule id {rule_id}",
-                    meta.code
-                );
-            }
-            for check_name in meta.codeclimate_check_names() {
-                assert!(
-                    check_name.starts_with("fallow/"),
-                    "result metadata code {} has unprefixed CodeClimate check name {check_name}",
-                    meta.code
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn codeclimate_result_exclusions_are_explicit() {
-        let expected = BTreeSet::from(["duplicate-prop-shape", "prop-drilling", "thin-wrapper"]);
-        let from_meta: BTreeSet<&str> = result_issue_metas()
-            .filter(|meta| meta.codeclimate_check_names().is_empty())
-            .map(|meta| meta.code)
-            .collect();
-        assert_eq!(expected, from_meta);
-    }
-
-    #[test]
-    fn codeclimate_result_codes_match_result_metadata() {
-        let result_codes: BTreeSet<&str> = result_issue_metas().map(|meta| meta.code).collect();
-        let codeclimate_codes: BTreeSet<&str> = CODECLIMATE_RESULT_CODES.iter().copied().collect();
-        assert!(codeclimate_codes.is_subset(&result_codes));
-    }
-
-    #[test]
-    fn ts_alias_policy_is_explicit() {
-        let aliases: BTreeSet<(&str, &str)> = result_issue_metas()
-            .filter_map(|meta| meta.ts_alias().map(|alias| (alias.name, alias.parent)))
-            .collect();
-
-        assert_eq!(
-            BTreeSet::from([
-                ("BoundaryViolation", "BoundaryViolationFinding"),
-                ("CircularDependency", "CircularDependencyFinding"),
-                ("DuplicateExport", "DuplicateExportFinding"),
-                ("EmptyCatalogGroup", "EmptyCatalogGroupFinding"),
-                (
-                    "MisconfiguredDependencyOverride",
-                    "MisconfiguredDependencyOverrideFinding",
-                ),
-                ("PrivateTypeLeak", "PrivateTypeLeakFinding"),
-                ("ReExportCycle", "ReExportCycleFinding"),
-                ("TestOnlyDependency", "TestOnlyDependencyFinding"),
-                ("TypeOnlyDependency", "TypeOnlyDependencyFinding"),
-                ("UnlistedDependency", "UnlistedDependencyFinding"),
-                (
-                    "UnresolvedCatalogReference",
-                    "UnresolvedCatalogReferenceFinding",
-                ),
-                ("UnresolvedImport", "UnresolvedImportFinding"),
-                ("UnusedCatalogEntry", "UnusedCatalogEntryFinding"),
-                ("UnusedDependency", "UnusedDependencyFinding",),
-                ("UnusedDependency", "UnusedDevDependencyFinding",),
-                ("UnusedDependency", "UnusedOptionalDependencyFinding",),
-                (
-                    "UnusedDependencyOverride",
-                    "UnusedDependencyOverrideFinding",
-                ),
-                ("UnusedExport", "UnusedExportFinding"),
-                ("UnusedFile", "UnusedFileFinding"),
-                ("UnusedMember", "UnusedClassMemberFinding"),
-                ("UnusedMember", "UnusedEnumMemberFinding"),
-                ("UnusedMember", "UnusedStoreMemberFinding"),
-            ]),
-            aliases
-        );
     }
 
     #[test]
